Set SQLAlchemy 'create_tables' on app startup
The Flask docs don't seem to go into great depths on how this is
to be used or if there are any implications to having an app run
this every time on startup. Some quick searching didn't reveal any
'do not do this' mentions in documentation or really anywhere, and
in practice it works fine.

A better alternative may be to setup a migration system with the
SQLAlchemy ORM, and only create/change db data with explict calls
to the migration system.