Added support for passing Authorizers to Minid Client
If someone wants to mint minids, but not do a Native Authorization
Flow. This supports other use-cases, like portals which may include
the minid scope in a 3-legged-auth flow and pass the authorizer
directly to the Minid Client. The Minid Client explicitly does not
automatically login via Native Auth to avoid possibly tying up a
thread in a non-native app context.